+++ Model/functions.py
@@ -7,18 +7,6 @@ import random as r
 import backend
 
 
-def info(simulation):
-    """ records the beginning of the step in real time and
-        prints the current step/number of cells
-    """
-    # records when the step begins, used for measuring efficiency
-    simulation.step_start = time.perf_counter()
-
-    # prints the current step number and the number of cells
-    print("Step: " + str(simulation.current_step))
-    print("Number of cells: " + str(simulation.number_cells))
-
-
 @backend.record_time
 def cell_death(simulation):
     """ marks the cell for removal if it meets
