+++ Model/templates/general.txt
@@ -8,13 +8,13 @@ What will the final step number be? Use for when you are beginning a new simulat
 [ 10 ]
 
 How many NANOG high cells to begin the simulation? Ex. 1000
-[ 100 ]
+[ 10000 ]
 
 How many GATA6 high cells to begin the simulation? Ex. 1000
 [ 0 ]
 
 What are the dimensions of the space in meters? Order as (x, y, z). Ex. (0.001, 0.001, 0)
-[ (0.002, 0.002, 0) ]
+[ (0.004, 0.004, 0) ]
 
 The time has come. Execute Order Sixty-Six? Ex. True
 [ True ]
