+++ Model/templates/imaging.txt
@@ -6,7 +6,7 @@ Do you want an image produced at each step and a video at the end of the simulat
 
 What is the image width in pixels? Currently imaging is supported for a 2D space, though a 3D space will yield
 a bird's eye view of the space. Image slicing/3D images to come... Ex. 2000
-[ 2000 ]
+[ 4000 ]
 
 How many frames per second of the output video that collects all step images? Ex. 1.0
 [ 6 ]
@@ -16,5 +16,5 @@ for coloring based on GATA6/NANOG levels? Ex. False
 [ True ]
 
 Include the FGF4 gradient color-map next to the image of the cells? Ex. True
-[ True ]
+[ False ]
 ------------------------------------------------------------------------------------------------------------------------
